[ci skip]

+++ src/compmake_plugins/backend_pmake/parmake_job2_imp.py
@@ -1,9 +1,11 @@
 import os
 import sys
-import traceback
+import time
 from contextlib import contextmanager
+from dataclasses import dataclass
+from multiprocessing import Queue
 from queue import Full
-from typing import Iterator, cast
+from typing import Any, Iterator, cast
 
 from compmake import (
     CMJobID,
@@ -41,7 +43,17 @@ def sanitize_for_filename(x0: str) -> str:
     return x
 
 
-async def parmake_job2(sti: SyncTaskInterface, args: tuple[CMJobID, DirPath, str, bool, DirPath]) -> ResultDict:
+@dataclass
+class ParmakeJobResult:
+    rd: ResultDict
+    time_total: float
+    time_comp: float
+    time_other: float
+
+
+async def parmake_job2(
+    sti: SyncTaskInterface, args: tuple[CMJobID, DirPath, str, bool, DirPath, "Queue[Any]"]
+) -> ParmakeJobResult:
     """
     args = tuple job_id, context, queue_name, show_events
 
@@ -51,8 +63,9 @@ async def parmake_job2(sti: SyncTaskInterface, args: tuple[CMJobID, DirPath, str
     because it might contain a Promise.
 
     """
-    job_id, basepath, event_queue_name, show_output, logdir = args
+    job_id, basepath, event_queue_name, show_output, logdir, event_queue = args
 
+    parmake_job2_started = time.time()
     ti: TimeInfo
     sanitized = sanitize_for_filename(job_id)
     mkdirs_thread_safe(logdir)
@@ -69,14 +82,16 @@ async def parmake_job2(sti: SyncTaskInterface, args: tuple[CMJobID, DirPath, str
     DEBUG_LOG = CompmakeConstants.debug_parmake_log
     if DEBUG_LOG:
         ZLogger.enable_simple = True
-    with timeit_wall(job_id) as ti, redirect_std(stdout_fn, stderr_fn, skip=DEBUG_LOG):
+    skip = DEBUG_LOG or True
+    with timeit_wall(job_id) as ti, redirect_std(stdout_fn, stderr_fn, skip=skip):
         sys.stderr.write(f"parmake_job2 {job_id} started\n")
+        sys.stderr.flush()
         check_isinstance(job_id, str)
         check_isinstance(event_queue_name, str)
-        from .pmake_manager import PmakeManager
+        # from .pmake_manager import PmakeManager
 
         # logger.info(f"queues: {PmakeManager.queues}")
-        event_queue = PmakeManager.queues[event_queue_name]
+        # event_queue = PmakeManager.queues[event_queue_name]
 
         async with MyAsyncExitStack(sti) as AES:
             with ti.timeit("contextinit"):
@@ -129,9 +144,17 @@ async def parmake_job2(sti: SyncTaskInterface, args: tuple[CMJobID, DirPath, str
                 publish(context0, "worker-status", job_id=job_id, status="connected")
 
                 with ti.timeit("make") as tisub:
-                    sys.stderr.write("make() started\n")
-                    res: MakeResult = await make(sti, job_id, context=context0, ti=tisub)
-                    sys.stderr.write("make() finished\n")
+                    sys.stderr.write(f"make() {job_id} started\n")
+                    sys.stderr.flush()
+                    try:
+                        res: MakeResult = await make(sti, job_id, context=context0, ti=tisub)
+                    except:
+                        sys.stderr.write(f"make() {job_id} failed\n")
+                        sys.stderr.flush()
+                        raise
+                    else:
+                        sys.stderr.write(f"make() {job_id} finished\n")
+                        sys.stderr.flush()
 
                 publish(context0, "worker-status", job_id=job_id, status="ended")
                 # r2: OKResult
@@ -142,29 +165,39 @@ async def parmake_job2(sti: SyncTaskInterface, args: tuple[CMJobID, DirPath, str
                 #     "deleted_jobs": list(res["deleted_jobs"]),
                 # }
                 res["user_object"] = None
-                result_dict_check(res)
+                if __debug__:
+                    result_dict_check(res)
                 res["ti"] = ti
-                return res
+
+                time_total = time.time() - parmake_job2_started
+                time_comp = res["time_comp"]
+                time_other = time_total - time_comp
+                return ParmakeJobResult(res, time_total=time_total, time_comp=time_comp, time_other=time_other)
 
             except KeyboardInterrupt:
                 msg = "KeyboardInterrupt should be captured by make() (inside Job.compute())"
                 raise AssertionError(msg)
             except JobInterrupted:
                 sys.stderr.write("job interrupted\n")
+                sys.stderr.flush()
                 publish(context0, "worker-status", job_id=job_id, status="interrupted")
                 raise
             except JobFailed:
                 sys.stderr.write("job failed\n")
+                sys.stderr.flush()
                 raise
             except BaseException:
                 sys.stderr.write("job exception\n")
+                sys.stderr.flush()
                 # XXX
                 raise
             except:
                 sys.stderr.write(f"parmake_job2 {job_id} another exception\n")
+                sys.stderr.flush()
                 raise
             finally:
                 sys.stderr.write(f"parmake_job2 {job_id} finished\n")
+                sys.stderr.flush()
                 publish(context0, "worker-status", job_id=job_id, status="cleanup")
                 setproctitle(f"compmake:{job_id}:done")
 
@@ -174,43 +207,46 @@ def redirect_std(stdout_fn: str, stderr_fn: str, skip: bool) -> Iterator[None]:
     if skip:
         yield
         return
-    sys.stdout.write(f"Activating stdout -> {stdout_fn}.\n")
-    sys.stderr.write(f"Activating stderr -> {stderr_fn}.\n")
+    sys.stdout.write(f"Activating stdout -> {stdout_fn}\n")
+    sys.stderr.write(f"Activating stderr -> {stderr_fn}\n")
 
-    new_stdout = open(stdout_fn, "w", buffering=1)  # 1 = line buffered
-    new_stderr = open(stderr_fn, "w", buffering=1)  # 1 = line buffered
+    with open(stdout_fn, "w", buffering=1) as new_stdout, open(stderr_fn, "w", buffering=1) as new_stderr:
 
-    old_stdout = sys.stdout
-    old_stderr = sys.stderr
-    sys.stdout = new_stdout
-    sys.stderr = new_stderr
-    resolution = "?"
-    try:
-        yield
-        resolution = "peaceful"
-    except:
-        resolution = "exception"
+        old_stdout = sys.stdout
+        old_stderr = sys.stderr
+        sys.stdout = new_stdout
+        sys.stderr = new_stderr
+        resolution = "?"
         try:
-            s = traceback.format_exc()
+            yield
+            resolution = "peaceful"
         except:
-            s = "Could not print traceback."
-        sys.stderr.write(s)
-        raise
-    finally:
-        # new_stderr.write(f"Closing stderr ({resolution=}).\n")
-        # new_stdout.write(f"Closing stdout ({resolution=}).\n")
-        sys.stdout = old_stdout
-        sys.stderr = old_stderr
-        new_stdout.flush()
-        new_stdout.close()
+            resolution = "exception"
 
-        new_stderr.flush()
-        new_stderr.close()
+            # try:
+            #     s = traceback.format_exc()
+            # except:
+            #     s = "Could not print traceback."
+            # sys.stderr.write(s)
+            raise
+        finally:
+            new_stderr.write(f"Closing stderr ({resolution=}).\n")
+            old_stdout.write(f"Now back to stderr ({resolution=}).\n")
+            # new_stdout.write(f"Closing stdout ({resolution=}).\n")
+            sys.stdout = old_stdout
+            sys.stderr = old_stderr
+            # new_stdout.flush()
+            # new_stdout.close()
+            #
+            # new_stderr.flush()
+            # new_stderr.close()
 
-        # sys.stdout.write(f"Recovering from {stdout_fn}.\n")
-        # sys.stderr.write(f"Recovering from {stderr_fn}.\n")
+            # sys.stdout.write(f"Recovering from {stdout_fn}.\n")
+            # sys.stderr.write(f"Recovering from {stderr_fn}.\n")
 
-        # delete the files if they are empty
+            # delete the files if they are empty
+
+    if False:
         try:
             if os.path.getsize(stdout_fn) == 0:
                 os.remove(stdout_fn)
